% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-simulated.R
\docType{data}
\name{simulated_pop}
\alias{simulated_pop}
\title{Simulated Benchmark Population by Origin}
\format{
A data frame with columns:
\describe{
\item{origin}{Origin area name.}
\item{population}{Benchmark resident population at origin.}
}
}
\usage{
simulated_pop
}
\description{
Simulated benchmark population counts by origin area.
}
\keyword{datasets}
