% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debiasR-package.R
\docType{package}
\name{debiasR-package}
\alias{debiasR}
\alias{debiasR-package}
\title{debiasR: Bias Correction for Mobile-Phone-Derived Mobility Data}
\description{
Tools to correct biases in origin-destination mobility data derived from mobile phone sources. Implements a range of bias-adjustment methods.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/de-bias/debiasR}
  \item Report bugs at \url{https://github.com/de-bias/debiasR/issues}
}

}
\author{
\strong{Maintainer}: Francisco Rowe \email{fcorowe@liverpool.ac.uk}

Authors:
\itemize{
  \item Carmen Cabrera \email{c.cabrera@liverpool.ac.uk}
}

}
\keyword{internal}
